Breaking change: only include images from app/assets/images
Only checking if the path includes 'images' isn't safe, because if there
is 'images' anywhere else in the path suddenly everything is an "image".
Also, it tries to include itself in itself too, because it's an "image"
too.

But this also means that now images from other dependencies aren't
included in the list anymore. But I didn't really find another way of
safely detecting only images.